Announce service IP addresses to the EGW
This announcement happens after an IP address has been allocated and
it tells the EGW about the service and the IP address.

I kept this operation separate from the IP address assignment just in
case we want to use a third-party IPAM.  This could also work with
pre-defined address pools a la MetalLB's allocator, but we'd need to
make that allocator a little more modular.