Simplify speaker configuration
The allocator's job is to allocate addresses and the speaker's job is
to announce them so it doesn't make sense to have the speakers trying
to figure out if the addresses are valid, the allocator should do
that.

At the moment the speaker controller doesn't have any configuration at
all, but the config passes through it in case we find that it needs
it.