Add code generator for client-side CRD
Based on kubernetes/kubernetes/staging/src/k8s.io/sample-controller

It's kind of a mess and looks like it was never upgraded to know about
modules, so it assumes/requires that it's running in the "GOPATH"
tree, which I haven't had a reason other than this to do.  It also
assumes a more recent version of some k8s API or other than metallb
was using so we'll probably need to upgrade some of the code in the
metallb k8s package.