Represent pools as ranges of addresses instead of groups of subnets
This might be mierreneuken but it has always bugged me that pools of
addresses were represented as lists of subnets.  I think that the code
is cleaner using two addresses to represent a range.

This commit introduces a "Pool" struct that manages one pool of
addresses.  The goal is that Pool can become an interface and then we
can have multiple implementations like Netbox, Acnodal, and local.