Massive change: merge egw and local pool allocators
We now have only one allocator process that can allocate addresses
from local pools and an Acnodal EGW based on its configuration.
Actually, since the EGW is roughly Netbox-compatible it should work
with a Netbox instance, too.

This involved a *lot* of refactoring, and also a lot of ripping code
out. The refactoring mostly involved wrapping the different allocation
strategies in an interface so the allocator doesn't have to know how
they work.

The code deletions mostly were around a change in approach: once
allocated, we will not remove an IP address from a service unless that
service is being deleted. Changing addresses creates instability and
we will only do it when the user explicitly asks us to. This also
simplifies the code significantly since there was a *lot* of code
devoted to undoing and redoing things when the configuration changed.