Update k8s.go to remove unused code and look more like the k8s sample controller
There were some "if" checks for conditions that can never happen.
Also, clientcmd.BuildConfigFromFlags can be used inside and outside of
the cluster so we don't need to have two cases.

It will likely be helpful when newcomers look at the code if it more
closely resembles the k8s sample controller.