Merge both announcers into one
At the moment the only one that's active is the local announcer.
We'll add configuration to determine which (or both) announces.

This is easier, though, for the users since they won't need to decide
which manifest to apply: there's only one.

In order to make this change I also pushed the "check local" and
election functionality down into the local announcer since we don't
need it for the acnodal announcer.