Enable Acnodal EGW service announcements
The code to announce services to the Acnodal EGW had gone missing when
we refactored the project so it needed to be added back.

While I did that I also moved some of the code so there's no longer an
internal/config package.  Our strategy used to be to parse a configmap
into internal objects and pass those around.  Now that we've got
custom resources we can pass those around and let each component
decide how to map them onto internal data structures.