Advertise non-local service addresses on the "dummy" interface
For local addresses we want to add the service address to one of our
nodes' default interfaces.  This tells Linux to respond to ARP
requests which attracts traffic to the cluster.  For non-local service
addresses, though, we can't add them to a physical interface so
instead we add them to a "dummy" interface which will not respond to
ARP requests.  Tools like routers will pick up the address and
advertise routes to the node.