Add ExternalTrafficPolicy for non-local addresses
We want to attract traffic to a node (i.e., add an address to the
"dummy" interface) only if that traffic will end up at a pod.  If the
ExternalTrafficPolicy is "Local" then we can only attract traffic to
pods that are running active endpoints since otherwise kubeproxy won't
forward to other nodes.  If the policy is "Cluster" then we can
advertise on all nodes because kubeproxy will forward.