Simplify EGW protocol
Now that we've got support for Netbox as an IPAM we don't need to keep
Netbox compatibility in the EGW. This change simplifies the protocol
substantially; it removes explicit IPAM address allocation and makes
it implicit when announcing a new LoadBalancer. The key changes that
allowed this to work were the previous few commits that let us pass a
v1.Service object into the pools.