Program service address differently for local vs EGW allocations
We program the service differently depending on where the address
came from.

If it's a locally-allocated address then we add it to the
LoadBalancerIngress.IP field which in ipvs mode will cause
kube-proxy to add the address to the kube-ipvs0 bridge
interface. This is needed for proper packet forwarding.

If the address was allocated from the Acnodal EGW, though, then
it should not be added to kube-ipvs0 because that would interfere
with proper packet forwarding to the endpoints after the packets
have been decapsulated. To do that we add the address to the
LoadBalancerIngress.Hostname field. This isn't documented well
but it's also done by cloud providers.

More info: https://github.com/kubernetes/kubernetes/pull/79976