Announce all endpoints, not just the first one
When there were more than one endpoint on a service we would only
check the first one and if it wasn't on our node we'd return instead
of checking the other endpoints. Now we check all of them so if our
endpoint isn't the first one we'll still announce it.