Support multiple EGW tunnels
Our current plan is to run a GUE tunnel to each client cluster node
for each service. In the case where there are multiple endpoints from
the same service on a single node we'll carry all of the traffic for
those endpoints over a single tunnel.

This requires making the Status struct more complex - it is now a map
so we can store EGW-end tunnel endpoint data for multiple client
nodes.

The announcers are a little more complicated but not much, and they're
more robust as they won't fail if an endpoint has already been
announced.

Because the announcers are getting info back from the EGW we don't
need to store as much data in annotations.