Speed up iprange calculations
The code to find the last IP address in a range was laughably bad (mea
culpa) and very slow. Now we use a third-party library to do that work
for us. One consequence is that some ranges can have both an IPV6 and
an IPV4 so we need to be more robust about comparing IPs and handle
the case where they're not from the same family, but they are
comparable.