Check pod/member sync and try to repair if needed
When we conduct an election we now check to see that there are the
same number of members in the memberlist as lbnodeagent pods in the
cluster. There should be. If there aren't, then something has gone
wrong and we try to recover by withdrawing from the memberlist and
rejoining.