Bind memberlist to a specific IP (not 0.0.0.0)
By using "0.0.0.0" we forced the memberlist package to decide which IP
address to use, and in some cases that decision would change over
time, which would cause a member's IP address to change, which would
confuse memberlist and cause instability.

Now we pick a specific address which makes memberlist stable even if
addresses get added or removed from the default interface.