Try harder to delete service IPs on nodes that aren't announcing
We've seen cases, maybe caused by race conditions, where we wanted to
withdraw an announced address but the service wasn't in our
announcement database. Now we provide the address as well as the
service name so we can still delete the address even if it wasn't in
the service db.