Release and re-allocate if desired pool doesn't match actual
We didn't handle the case where the user edited a service and added or
changed the purelb.io/service-group annotation. Now we release the
address back to the pool from which it was allocated, and allocate a
new address from the requested pool.

Note that this should only be used in a development environment. In
production you should always create a new service, transition to it,
and delete the old one only when you're sure that it's no longer
handling traffic.