Fix a couple of local address pool shared allocation corner cases
We had a couple of bugs in our localpool address sharing. In one case,
we were deleting the sharing key too soon, i.e., before all of the
users of the address had been unassigned. This allowed services to
"steal" an address that they should have been blocked from.

In another case, we were deleting the wrong service from the
portsInUse struct which had a similar effect: services that shouldn't
have been able to use an IP address were able to.