Handle re-allocation failures
We weren't properly checking the error return values when we notified
the address pools about an existing allocation. This let some corner
cases slip through, for example, when the user edited a service that
already had an IP address. The user could edit things like the service
port and it wouldn't be checked. Now it is.