Reduce unwanted address allocs by updating status first then spec
Update the status first. We used to update the spec and annotations
first but that caused us to attempt to allocate twice because the
spec/annotation update caused an event but the service address (which
is part of the status) wasn't set yet.

Now that we update the status first the event that it generates
doesn't cause another allocation attempt because the controller sees
that the service already has an address.