Track LoadBalancerIngress instead of IP address
We made a change earlier where we set the ingress address differently
in different cases. In some cases we set an IP address and in other
cases we set a HostName, but we were caching the IP address only. Now
we cache the LoadBalancerIngress object which has both the address and
hostname.