Don't hide errors when addresses are unassigned
MetalLB assumed that unassigning an IP address could never fail, which
is true in its case since it only handles local address pools. We work
with remote address pools like Netbox and Acnodal's EPIC, so we need
to percolate errors up from the pool to the controller so the
controller can retry.