Handle initialization of svcs with Ingress Hostnames
We need to program the Service Ingress differently depending on
whether the Ingress address is a local address or an Acnodal EPIC
address. In the EPIC case we program the Ingress Hostname, not the
Ingress IP.

When the allocator starts up it scans all services to discover which
addresses had been previously allocated.  We weren't properly handling
services whose Ingress had been assigned a Hostname instead of an
IP. Those services worked, but because the allocator didn't know about
them it would fail to notify the EPIC if they were deleted.

This commit includes some refactoring to move allocation code from
the controller to the allocator. I think it makes the separation of
concerns cleaner because now more code that relates to allocation is
in the allocator.