Switch from our ad-hoc purelb-complete.yaml to kustomize
This was driven by a couple of goals. First, we want to generate
versioned artifacts. We want our users to have precise control over
what they run, so we want to give them manifests that load specific
versions and won't drift over time.

We were tagging images and manifests with just "ci" which isn't very
useful. Now we tag each build depending on what triggered the build.
If the build was triggered by a tag (which is usually the case for
releases) we use that as the build tag. If the build was triggered by
a commit on a branch (which is common during development) the build
tag will be ${CI_COMMIT_REF_NAME}-${CI_COMMIT_SHORT_SHA} which works
out to something like "gitbranch-24e0a368" which tells you which
branch and which specific git commit was used to build the image.

The generated deployment manifests now load specific docker images so
you don't need to worry about your images changing without your
explicitly telling them to do so. If you apply one of these manifests,
PureLB won't change until you apply a different manifest.

We also build per-branch "latest" manifests with names like
"${CI_COMMIT_REF_NAME}-latest". These will load the most recent images
on that branch. They can be useful in development when you want to
track a branch and update to new commits by deleting the PureLB pods,
but they are not recommended for production systems.

Second, we want the build to work in different environments,
potentially even with different project names. We now have a
${PROJECT} variable in the Makefile which the CI config sets to the
name of the project.

Kustomize is still pretty "green" so I'd recommend installing a recent
version. I'm confident that "kubectl -k" will not work because the
version of kustomize that's baked into kubectl is too old. The CI
config currently installs v4.0.1 and it's not certain that any older
version will work properly.