Manage Acnodal EPIC "clusters"
A "cluster" in this case corresponds to a k8s client cluster. A single
EPIC load balancer can serve multiple back-end k8s clusters so each
needs to announce itself to EPIC. To make it easy to use, the first
cluster to announce will also create the LB, and subsequent clusters
will add themselves to the existing LB. It works the same way in
reverse, too.