Avoid retry loops when something goes wrong during service cleanup
We were getting stuck retrying many times. This is an error condition,
for example, when a purelb client creates a service on the Acnodal
EPIC and then someone deletes the service by hand. PureLB should try
to clean up but if something goes wrong it doesn't need to keep
retrying, it should move on.