Refactor the use of the CNI_INTERFACE to prepare for configurability
We need to configure where we attach the PFC to the CNI interface
because different CNIs configure the network differently. This change
reduces the number of places where we use the CNI_INTERFACE const, and
ensures that all of those places are also places where we can access
the configuration CRs so it should make it easy to switch.