Make PFC network attachment configurable
We used to use a lot of hard-coded values to attach our PFC to their
network interfaces. Now those values are configured in the LBNodeAgent
CR so we have more flexibility in how we set up the PFC. This will let
us work with a variety of CNIs, not just Flannel.