Generate several deployment descriptors
We need deployment descriptors for a few different use cases including
the PureLB that runs in the EPIC cluster, client-side PureLB clusters
that talk to EPIC, and client-side PureLB clusters that advertise
addresses locally.

This change allows us to version the manifests by project, version,
and "configuration base" which is just the directory in which we run
kustomize.

The CI script generates a few manifests and either includes them in
the job artifacts (for branch/commit builds) or uploads them to the
project package repo (for tag builds).