Use configured interface for encap attachment instead of "cni0"
https://gitlab.com/acnodal/ego/-/issues/36

With this change we now treat the encap and decap interfaces the same
from a configuration perspective. If either is specified then we use
the specified interface, otherwise we look for the interface with the
lowest cost default route.

This commit also improves our cleanup. It's still a little sketchy
since when we're cleaning up we don't necessarily have the old
configuration so we might be cleaning up something different that we
previously set up, but it's the best we can do without caching the
full configuration.