Implement a "default" service group
We need to handle the case where we're running in a cloud provider's
context and the user wants the cloud provider's controller to own an
LB. In other words, we need a way for the user to tell us to ignore an
LB. Before we would grab an address from any available pool but that
approach proved too aggressive. Now if there's no
DesiredGroupAnnotation we try to allocate from a service group named
"default" but ignore the service if there is no "default" service
group.

This works with the ServiceLBClass that we just implemented. Users
running PureLB in cloud providers should not configure a "default"
service group. This will ensure that PureLB will not allocate
addresses unless the user specifically requests it to, on a
service-by-service basis. In other words, the cloud provider's
orchestrator will be the default, not PureLB.

Users running in an enterprise environment (like their own data
centers) should configure a "default" service group. PureLB will then
be the default orchestrator and will allocate addresses for all
services unless told not to by marking the service with a different
orchestrator's ServiceLBClass.