Use UBI minimal base Docker images
"The Universal Base Image Minimal is a stripped down image that uses
microdnf as a package manager."

https://catalog.redhat.com/software/containers/ubi8-minimal/5c64772edd19c77a158ea216

We were using Alpine for the allocator and Ubuntu for the node
agent. This is easier since we now have a single Dockerfile.

One side-effect of this change is that we no longer support
linux/arm/v7 binaries, i.e., 32-bit Raspberry Pis. You will need at
minimum an RPi 3 running 64-bit Raspberry Pi OS to use PureLB. We
recommend at least an RPi 4, though.