Don't Unassign addresses when Assign fails
This can cause damage in a specific case: the allocator restarts and
the startup-time service announcements arrive before the startup-time
CR announcements. This means that when the allocator tells the pool to
Assign() the pool won't know about the service group yet so the
assignment will fail.

We were Unassigning the address in that case but we shouldn't. We
should leave it alone because in most cases the CR event will arrive
soon after and that will trigger a rescan of all services. Even if it
doesn't we don't want to modify the Service, just leave it alone.