Implement a helm chart using our kustomize resources
This helm chart is pretty naive. I looked at some of Bitnami's
charts (including MetalLB's) and they were a curly-bracket soup of
weird golang templating. I don't think that we need that much
flexibility and I like for the yaml to look somewhat like yaml so I'm
keeping the set of things that you can customize limited for now.

As we find things that people *actually* want to customize we can
convert those to templates driven by variables.