Poll only for service changes, not config or endpoints
We don't need to poll config changes because they all happen locally
to this cluster so we'll always be informed of events.

We want to poll EPIC to find out about changes to the service such as
new proxy pods launching, but we only need to poll it once so we set
the svcIndexer to poll but not the epIndexer.