Upgrade k8s.io modules to v0.21.0
We need to do this to get access to the Service.Spec.loadBalancerClass
field.

I'm only upgrading to v0.21 and not all the way to v0.22 because v0.22
requires go v1.16 and I don't feel like upgrading my compiler yet. I
will probably have to do so before long, though.