Add a check for svc.Spec.LoadBalancerClass
This feature allows a user to have multiple load balancer
orchestrators installed at the same time and choose which one handles
each service.

If Service.Spec.LoadBalancerClass is not set then PureLB will handle
that service. This is the default. If Service.Spec.LoadBalancerClass
is set, though, and it's not "purelb.io/purelb" then PureLB will
ignore that service because some other orchestrator will handle it.