Implement a default ServiceGroup
Before, we would fail to allocate an IP address if there was no
"purelb.io/service-group" annotation. Now if there's no annotation we
try to allocate from a service group named "default".

This works with the ServiceLBClass that we just implemented. Users
running PureLB in cloud providers should not configure a "default"
service group. This will ensure that PureLB will not allocate
addresses unless the user specifically requests it to, on a
service-by-service basis. In other words, the cloud provider's
orchestrator will be the default, not PureLB.

Users running in an ad-hoc environment (like their own data centers)
should configure a "default" service group. PureLB will then be the
default orchestrator and will address allocation for all services
unless told not to by marking a service with a different
orchestrator's ServiceLBClass.