Increase PFC session GC from 30s to 1h
The GC is remarkably stupid - it simply waits until a session is a
certain age and then blindly kills it. The problem is that Envoy tries
to keep connections to its upstream endpoints open, so it thinks that
it has an open connection but the sweep_sessions has killed the
connection so the next time Envoy tries to make a request it fails.

The real answer is to instrument the PFC so that it won't kill active
sessions but in the mean time I'm going to make the timeout 1 hour
which should work OK for our purposes until we can fix it for real.