Use TunnelID instead of GroupID/ServiceID
We found a case where gid/sid wasn't unique enough: when Envoy pods
belonging to a load balancer but running on two different EPIC nodes
set up tunnels to the same PureLB node then the TABLE-ENCAP didn't
work because it was keyed by gid/sid.

To work around this with the same PFC we're splitting the TunnelID in
half and stuffing it into the gid and sid fields. This gives us enough
uniqueness to get the packets back to the EPIC node from which they
were sent.