Add a flag to configure whether PureLB is the default announcer
PureLB now checks the DEFAULT_ANNOUNCER environment variable to
determine whether to handle services with no explicit
Spec.LoadBalancerClass. If the DEFAULT_ANNOUNCER env var is set to the
value "PureLB" then PureLB will handle services with no explicit
Spec.LoadBalancerClass. If it is set to any other value, or not set,
then PureLB will ignore services with no explicit
Spec.LoadBalancerClass.

This commit sets the value of DEFAULT_ANNOUNCER to "PureLB" so PureLB
will act as the default announcer unless the user changes it.