Refactor CI script
This refactor has two goals:

* reduce build times by increasing concurrency
* separate more "go-ish" tasks from more "docker-ish" tasks

We've reduced build times by running the "check" job in parallel with
the "image" jobs. The "check" job is usually the longest one so now we
don't have to wait for the images - they're usually finished before
the check is finished.

The manifest job did some docker stuff and some golang stuff, but that
makes setting up the image a little clumsy since you either need to
install docker in the golang image or vice versa. Now we split the
"manifest" job into two parts: one that uses the golang image, and one
that uses the docker image. This makes installation easier.