Use controller-gen to generate CRDs
These used to be hand-crafted but manually keeping the CRD manifests
in sync with the golang code gets boring. We now use
controller-gen (part of the controller-tools project) to generate the
CRDs for us.

controller-gen has less flexibility in terms of its output than
hand-crafting, so we've lost one feature from our CRDs: we can no
longer use the OpenAPI "oneOf" directive, so where the CRD used to
enforce that there was *either* a "local" service group *or* a
"netbox" service group, we now no longer have that check. Our code
checks for a "local" group first so if both are specified then the
user will get a local group and the netbox group will be ignored.

Because the generated CRD is more permissive than the hand-crafted one
we can install the new CRD over the old one and all of the resources
will be valid.