Refactor allocator and controller to move Ingress mgmt to allocator
We used to manage the Service.Status.Ingress at the controller level
but this complicated the interface between the controller and the
allocator because the Ingress needs to be programmed differently
depending on which pool the address came from, so the controller
needed to know more about how the pools work then it should.

Now the allocator manages the Service's Ingress. It's not perfect -
the pools should probably do this - but the interaction between the
controller and the allocator is cleaner.

I was able to lower-case some of the allocator's functions because now
they're only called from within the allocator. Granted, the controller
is in the same package so it has access to them, but it doesn't need
to call them.