Add IPV4 and IPV6 pools to ServiceGroupLocalSpec
This allows users to configure both IPV4 and IPV6 in a single
pool.

I left the top-level Subnet, Pool, and Aggregation parameters because
they're more convenient for users in a single-stack environment.

I considered using an array of ServiceGroupAddressPools instead of
explicit V4Pool and V6Pool but I think that the explicit approach is
easier to understand, if somewhat less flexible.