Have pools track their own allocations
I want to simplify the interface between the allocator and the
pools. This is a start - the allocator no longer passes the IP address
into the pool when releasing it. This means that the allocator no
longer needs to track the address but the pools do.

I wasn't able to completely remove the address tracking from the
allocator because it's used by the unit tests, but it's no longer used
in the allocator/pool interface.