Move Ingress address assignment into the Pools
We used to have the Pools provide addresses to the Allocator and then
the Allocator would assign the address to the Service. Now that we're
working to enable dual-stack, this gets messy. It seems cleaner to
have the Pools choose the addresses and assign them to the Service
Ingress. This simplifies the interface between the Allocator and the
Pools because the Allocator no longer needs to handle addresses.