Use the service's UID as the upstream cluster name
This has taken a while to get right, but hopefully we're
there. Multiple client-side Services can share a single EPIC load
balancer. We call each service a "cluster" because it becomes an Envoy
upstream cluster. It's been difficult to get the logic for adding and
removing clusters and LoadBalancers correct, because there are lots of
corner cases where you want to block or allow LB creation and
deletion.

One problem we had was how to name the clusters. The name needs to be
unique but also short enough to be a k8s label value. I now
think (after trying a few incorrect approaches) that the simplest
approach is to use the object ID that k8s assigns to the client-side
Service object. This is unique in "space and time" according to the
docs. It's also a fixed size so it won't cause problems if we use it
as a label value.

A side benefit of this approach is that it gets us closer to upstream
PureLB. It doesn't need to track a cluster ID at all, and now our
branch doesn't either because it can just grab the Service object ID
when it needs it.