Switch default image tag to "main-latest"
We changed our build process to include the branch name in the image
tag, but we didn't change the default manifest so if a user followed
the instructions in the README they'd end up with very old images.

This changes the image tag to "main-latest" which will be relatively
fresh but volatile over time.