Use k8s event messages to indicate ServiceGroup parse results
Previously if something went wrong when parsing a ServiceGroup it
wasn't immediately obvious to the user. We don't have a validating
webhook so we won't reject an invalid SG, but we won't use it, either.
Users had to scan the allocator logs looking for errors.

Now we pass the k8s.ServiceEvent interface into the allocator so it
can attach event messages to ServiceGroup resources as it parses
them. This lets users "kubctl describe" the SG to check whether it was
parsed successfully or not, and why.