Make lbnodeagent pod labeling configurable
This gives users more flexibility in assigning labels to
lbnodeagents.

The memberlist package that manages the election process needs to be
given a list of pods to bootstrap. We query for the lbnodeagents using
a set of labels, and then pass those pods' addresses to
memberlist. The label query used to be hard-coded but now it's passed
in to the lbnodeagent via the manifest, and can be managed in the helm
charts.

Closes #88
Closes #89