Warn instead of releasing if allocator thinks service has an address
We've seen cases where a service gets into a state where our allocator
has allocated an address for it but that address hasn't been written
into the service's Status yet. If, for example, the status update that
adds the Ingress fails. Warning is less brutal than releasing the
address, and less likely to cause secondary damage.