Use an ip from the decap intf instead of the node ip for tunnels
We were using the node interface as the address of this end of GUE
tunnels but that's not always what we want to use. Now we pick the
first address on the tunnel interface (which is either specified or a
default) in the same family as the IP address on the other end of the
tunnel.