Allow more than one announcement interface
The LBNodeAgent.Spec.Local CR used to allow the user to specify the
localint as "default" or the name of a single interface. This proved
too inflexible in some cases, for example where a host has two
different interfaces with disjoint networks and the user wants to
announce on both interfaces.

The CRD hasn't changed, but now LBNodeAgent.Spec.LocalInterface can be
either "default" or a regex. This allows the user to specify patterns
of interface names and the announcer will scan the interfaces that
match those patterns. If an interface matches the name regex and its
network contains the LB address then the address is added to the
interface. If no interface matches both criteria then the address is
added to the remote interface.

Regexs don't have to include wildcards so this change is
back-compatible since, for example, "enp5s0" is a regex that will
match the interface enp5s0.