Add shortNames for our CRDs
We used to have shortNames for our custom resources, so you could, for
example, run `kubectl get -n purelb sg` to list the service groups. At
some point (probably when we switched to generated CRDs) those
shortNames were lost.

This commit adds kubebuilder directives so "sg" and "sgs" will be
shortNames for ServiceGroup, and "lbna" and "lbnas" will be shortNames
for LBNodeAgent.