Fix a merge mistake where lbnodeagent had two config blocks
In the previous release two community members both added features to
the helm chart and I failed to see all of the merge points, so we
ended up with two LBNodeAgent sections in values.yaml.  One was called
"lbnodeagent" and the other was called "LBNodeAgent".

We need to pick one and it looks like the path of fewest changes is to
use "lbnodeagent" and remove "LBNodeAgent" so I'm going to go with
that.