Fix explicit assignment of IPV6 LB addresses
Issue #106

The user can assign an IP address to an LB with the
spec.loadbalancerIP field, but this didn't always work for IPV6
addresses. If the pool had both IPV4 and IPV6 ranges then Contains()
would return only whether the address was contained in the IPV4 range
and not check the IPV6 range. Now it checks both.