Add multiple address ranges to LocalPool
Given the global IPV4 address exhaustion many people running
bare-metal clusters will get their IP addresses in small chunks so it
will be convenient for them to assemble a few small chunks into a
single pool and let PureLB manage it.

This change is big but pretty straightforward. In many cases where we
had a single range we now loop over all of the the ranges in the pool.

The tricky part of this change was LocalPool.next(). I decided to keep
the ranges in the same order that the user specifies so if the ranges
are out of order then next() might return a *lower* address than
expected. This can happen when one range becomes exhausted and the
next range has lower addresses than the previous range.

See also https://gitlab.com/purelb/purelb/-/issues/109