Don't let bad address pools block good ones
We parse all of the known ServiceGroups every time one of them
changes. That's OK (if a little inefficient) but we used to bail out
of the parsing if any of the ranges had any error. This meant that if
you had two SGs (one good and one bad) you couldn't use either one.

Now we mark the bad SGs but use the good ones.

Issue #75