Import ip-masq-agent's code
https://github.com/kubernetes-sigs/ip-masq-agent/ looks like a good
model for what we need to do with our egress feature and it's very
small and licensed under the same Apache license as PureLB.

The chaos in go.mod needs a little explanation. ip-masq-agent uses
code in Kubernetes that hasn't been exported into a subproject so it
requires k8s.io/kubernetes. The go.mod in that project, though, lists
all of its dependencies as "v0.0.0" (not sure why) which causes "go
get" to barf. There's a long discussion at
https://github.com/kubernetes/kubernetes/issues/79384 that has various
workarounds and the one I used was the script in
https://github.com/kubernetes/kubernetes/issues/79384#issuecomment-521493597
that generates the replacements needed. I saved a copy of that script
as hack/generate-k8s-requires .

This import required that we upgrade our golang version. We were on
1.15 which doesn't have the io/fs package so I upgraded to the latest
and greatest.