Build one container image instead of two
I've been wanting to do this for a while. It simplifies the build
infrastructure, speeds up CI builds, and should be a little easier at
runtime, too, since the cluster only has to download one image instead
of two.

Another benefit is that now the Dockerfile is in the usual place which
should make it easier for developers.