Add a simple routing table allocator for egress tables
Routing tables have IDs which are usually mapped to names in
/etc/iproute2/rt_tables. In our case we need a table per service so
we'll require that the user find a range of table IDs. The current
default is 500 which might be OK for most users.

The allocator is lossy: it simply bumps the counter every time it sees
a new service. This should be fine for production use which is
relatively stable.