Send a gratuitous ARP message when adding an address to an interface
The Linux network interface doesn't send a GARP when you add an
address to an interface, but we want to do that since we add addresses
to interfaces when an address fails over from one node to another.

I used the MetalLB code as a base since they do all sorts of ARP
stuff.