Allow Local ExternalTrafficPolicy with configuration override
We don't think that running with ExternalTrafficPolicy set to Local is
a good idea, because k8s won't forward traffic to nodes that are
running active service endpoints. This means that we might announce
from a node that has no endpoint and then all traffic would be lost.

At least one user is OK with that, though. They're running their
service on *every* node in their cluster so it should always have a
service endpoint regardless of which node we announce from. So we want
to support it, but only when the user indicates that they understand
what's going on.

To enable Local, add a "purelb.io/allow-local" annotation to the
service. The value of the annotation can be any string.