Add a "make scan" target that runs govulncheck
I saw a blog post about the govulncheck tool and it seems like running
it every now and then is a good idea so I added a "scan" target to the
Makefile. When I ran it it found a couple of vulnerabilities but one
of them I would need to upgrade the go major version to fix but I'm
not ready to do that yet. It will probably be fixed in a minor version
before too long.

The other vulnerability was in the "net" package which seems easy
enough to upgrade so I've done that, although it seems that it dragged
a few other upgrades in, too.