fixup! Add multiple address ranges to LocalPool
87603b32 didn't work - non-local addresses didn't get assigned because
the new multi-pool code didn't manage some of the functions that the
lbnodeagent needs.

This fix adds code to the pools so we can figure out which pool a
given address belongs to. We need to do this because the address is
allocated by a single process (the allocator) but then the address
needs to be announced by possibly many processes (the lbnodeagents).

See https://gitlab.com/purelb/purelb/-/issues/124

Signed-off-by: Toby Cabot <toby@acnodal.io>