Allow the addresses annotation to override desired-group
Previously we would fail if the user specified both the "addresses"
annotation and the "desired-group" annotation and they didn't agree
with one another, e.g., if the specified address wasn't in the
specified pool.

Now the "addresses" annotation completely overrides "desired-group" so
"desired-group" is ignored (with a warning) if "addresses" is
specified.

This feels a little more user-friendly.