Simplify interface between service and allocator
We'd like to have the inner workings of the Allocator not bleed into
its interface with the Service. Before, the Allocator returned a pool
name, and then the Service used it to do some housekeeping. Now the
Allocator does the housekeeping so the Service just needs to know if
there was an error or not.