Generate client-side code on the fly, using a specific generator version
Not sure why we did this, but we used to generate our client-side
interface code and then commit it to git. This is a recipe for trouble
long-term since we can silently lose the ability to generate if the
tools become obsolete.

Now we generate that code as part of the build process, and it's not
committed to git since it's not source code.

Until now, we had code that would scan for the code-generator script,
first in the vendor directory and then in a sibling directory to
purelb. This is useful if you need to make changes to the generator
itself, but it means that local builds depend on the contents of that
sibling working directory, which might have local edits that change
the generator output. Now we run a specific version of the generator
from where Go has installed it.

The generator itself changed pretty radically: it requires a different
package structure than it used to. The script entrypoint is also
different. You now run the generators by sourcing kube_codegen.sh and
calling bash functions that it creates.