Generate client-side code on the fly, using a specific generator version
Until now, we had code that would scan for the code-generator script,
first in the vendor directory and then in a sibling directory to
purelb. The output of the generator was then committed to git.

This change uses a specific version of the code generator (the one in
go.mod) and generates on the fly. This should allow for more
repeatable builds.